Added sticky top nav-bar + disclaimer
I made the top navigation bar sticky, meaning it will stick to the top of the page after scrolling past it.
Therefore I removed it from the heading div and made its own div.

I also added a disclaimer by the "What are NBT's" that the information was taken from the NBT website.